Add Unicode character escape \u{X+} to OCaml string literals.
The syntax of string literals is changed to add the \u{X+} escape
sequence which replaces the escape by the UTF-8 encoding of the
Unicode scalar value denoted by the hexadecimal number H made of one
to six lower or uppercase hexadecimal digits.

Unfortunately due to OCaml lax policy about escape sequences this
change can break program that trigger warning 14 on compilation. More
precisely programs which have literals with "\u" as a substring will
now either fail or the string literal will be silently compiled to
another representation in case the literal has a subsequence that
respects the \u{X+} syntax.