parmatch: [build_specialized_submatrices] returns the default matrix as well
This is a post merge cleanup. [build_specialized_submatrices] always builds the
default bmatrix as part of its work, but was previously only returning it when
there were no specialized matrices and when the [return_omega_group] parameter
was [true] (yes, this was ugly).
As it turns out, all the places calling this function also ended up building
the default matrix as a follow up.
The function now always returns the default matrix along with the (possibly
empty) list of specialized matrices.

As a consequence [build_default_matrix] became redundant and was removed.

Thanks to Gabriel for noticing this.