rec_check: remove the duplicate array_type_kind classifier, use Typeopt
A side-result of this change is that, because Typeopt classifies
parametric arrays in a more flexible way that the Rec_check code did
(when -no-flat-float-array is set, they are marked Paddrarray rather
than Pgenarray, to reflect the fact that not dynamic test for unboxing
is performed), the -no-flat-float-array compiler starts accepting
*more* recursive definition than before: basically, arrays at all
types act just like constructors.

Our testsuite had a case (letrec-disallowed/generic_arrays.ml) that
specifically check that this case of generic array is disallowed,
even under -no-flat-float-array. I changed the tests to reflect
the finer-grained criterion -- and converted them to expect-style
tests along the way.