[minor] clarify the implementation of Typedtree.pat_bound_idents_full
The previous implementation used global state and was making another
change more difficult than it had to be.

Note: I got rid of `rev_let_bound_idents` in the API because computing
it is no more efficient than `Fun.compose List.rev let_bound_idents`.

(Only the _full versions might have a more-efficient rev_ version,
and those are not exposed, which is the right choice:
                           a simpler API is worth a list reversal.)